stupid cast

+++ Projectiles/Magic/EternityHex.cs
@@ -151,7 +151,7 @@ namespace CalamityMod.Projectiles.Magic
                 {
                     int homerCount = 6;
                     int trueMeleeID = ModContent.ProjectileType<EternityHoming>();
-                    int trueMeleeDamage = Eternity.BaseDamage * player.MagicDamage();
+                    int trueMeleeDamage = (int)(Eternity.BaseDamage * player.MagicDamage());
                     float angleVariance = MathHelper.TwoPi / homerCount;
                     float spinOffsetAngle = MathHelper.Pi / (2f * homerCount);
                     Vector2 posVec = Vector2.UnitX.RotatedByRandom(MathHelper.TwoPi) * 10f;
